fix virtual interview (maybe)

+++ virtual_interview_analyze.py
@@ -22,15 +22,16 @@ logging.basicConfig(level=logging.INFO)
 
 def mp4_to_wav(mp4_filename):
     # Define file paths
-    mp3_filename = 'speech.mp3'
-    wav_filename = 'speech.wav'
+    mp3_filename = './speech.mp3'
+    wav_filename = './speech.wav'
 
     # Convert mp4 to mp3
+    # ffmpeg -i audio.wav -acodec libmp3lame audio.mp3
     command2mp3 = f"ffmpeg -i {mp4_filename} {mp3_filename}"
     result = subprocess.run(command2mp3, shell=True, capture_output=True, text=True)
-    if result.returncode != 0:
-        print(f"Error converting mp4 to mp3: {result.stderr}")
-        return
+    # if result.returncode != 0:
+    #     print(f"Error converting mp4 to mp3: {result.stderr}")
+    #     return
     
     # Convert mp3 to wav
     command2wav = f"ffmpeg -i {mp3_filename} {wav_filename}"
@@ -85,20 +86,32 @@ def video_prediction(img):
     return results
 
 def wav_prediction(file_path):
-    mp4_to_wav(file_path)
-    wav_path = "./speech.wav"
-    pred_emotion, pred_fluency, average_fluency = model.predict(wav_path, cutdur=2)
-    max_fluency = 2
-    min_fluency = 0 
-    normalized_fluency = (average_fluency - min_fluency) / (max_fluency - min_fluency) * 100
-    pred_emotion, pred_fluency, normalized_fluency
-    os.remove('./speech.mp3')
-    os.remove('./speech.wav')
-    return {
-        'emotion_prediction': pred_emotion,
-        'fluency_prediction': pred_fluency,
-        'pronunciation_score': normalized_fluency
-    }
+    try:
+        mp4_to_wav(file_path)
+        wav_path = "./speech.wav"
+        pred_emotion, pred_fluency, average_fluency = model.predict(wav_path, cutdur=2)
+        max_fluency = 2
+        min_fluency = 0 
+        normalized_fluency = (average_fluency - min_fluency) / (max_fluency - min_fluency) * 100
+        pred_emotion, pred_fluency, normalized_fluency
+        os.remove('./speech.mp3')
+        os.remove('./speech.wav')
+        return {
+            'emotion_prediction': pred_emotion,
+            'fluency_prediction': pred_fluency,
+            'pronunciation_score': normalized_fluency
+        }
+    except:
+        if os.path.exists('./speech.mp3'):
+            os.remove('./speech.mp3')
+        if os.path.exists('./speech.wav'):
+            os.remove('./speech.wav')
+        return {
+            'emotion_prediction': 'Voice not detected',
+            'fluency_prediction': 'Voice not detected',
+            'pronunciation_score': 'Voice not detected'
+        }
+    
     
 def prediction(video_path):
     cap = cv2.VideoCapture(video_path)
@@ -120,6 +133,12 @@ def prediction(video_path):
                 'emotion': pred['emotion']
             })
         frame_count += 1
+    print(len(video_pred))
+    if len(video_pred) > 7:
+        step = (len(video_pred)-1) / (7-1)
+        print(step)
+        video_pred = [video_pred[round(i*step)] for i in range(7)]
+            
     cap.release()
     cv2.destroyAllWindows()
     voice_pred = wav_prediction(video_path)
